feat: divided app.js in smaller files, added sessions to dbSetup
